feat: move bulletPoint to inputs common component
